requirements.txt
livekit-agents
python-dotenv
openai
deepgram-sdk
elevenlabs
cartesia 
